엑세스 로그 저장 비동기 통신 방식 및 배치 처리 적용

+++ src/main/java/com/project/bankassetor/model/response/AccountCreateResponse.java
@@ -13,7 +13,7 @@ public class AccountCreateResponse {
     // 계좌 유형 (예: SAVING, CHECKING)
     private String accountType;
     private long accountNumber;
-    private long balance;
+    private int balance;
 
     public static AccountCreateResponse of(Account account){
         return AccountCreateResponse.builder()
